requirements.txt
Flask
gunicorn
Flask-WTF
WTForms
Flask-Login
Flask-SQLAlchemy
email-validator
itsdangerous
Jinja2
Werkzeug
python-dotenv
pillow              
requests
flask-bcrypt           
flask-mail
flask-migrate
psycopg2-binary
